requirements.txt
streamlit
openai>=1.0.0
pandas
schedule
python-dotenv
praw
